Link against wayland libs using info from pkg-config
Right now the automake file is ignoring the input from pkg-config
entirely when linking compiled binaries against the wayland-client and
wayland-server libraries. -lwayland-server and -lwayland-client might
work on a good number of systems, however if any user tries to compile
this with their wayland libraries in a non-standard location then
compilation will fail even though that's a perfectly valid thing to do
with autotools.

So, use $(WAYLAND_LIBS) so that the linker flags we use are guaranteed
to always be correct.