Add pkg-config files for both libraries and protocol sources
This allows us to actually integrate this library and it's XML sources
into other build systems without having to pull in all of the files for
the project locally or keep them up to date. It also helps us avoid
having to figure out the linking arguments ourselves in build systems.