Fix race condition when creating a display
Some tests revealed a race condition happening when multiple threads
call wlEglGetPlatformDisplayExport() at the same time. The API lock is
released when roundtriping, which may lead to different threads messing
with the global display and device lists at the same time.

This change reworks wlEglGetPlatformDisplayExport() such that the race
condition is avoided.