Properly generate private/public protocols
The 'code' option for wayland-scanner was deprecated in favor of
'private-code' and 'public-code'.

This change updates the autotools build to properly use the new options
and updates the meson build to generate private-code by default and only
use public-code for public interfaces.

Additionally, the default visibility for the entire library is set
hidden with '-fvisibility=hidden'