Create per-thread wl_event_queues and use proxy wrappers
Our Wayland interaction model was written so that we would use a
separate per-display event queue. This was done to make sure responses
to requests originated from EGL were only handled by EGL, without
interfering with application requests or events. This approach is
suggested in Wayland's documentation:

  https://wayland.freedesktop.org/docs/html/apb.html#Client-classwl__display

However, on a multi-threaded environment, using the same queue for all
threads may cause similar interaction issues, even when those threads
are internal to the EGL Wayland library. An example of that in our
implementation is when using EGL_NV_stream_fifo_synchronous, where we
employ a separate damage thread to monitor when images are finished and
notify the compositor. In that scenario, we would use a separate event
queue for the damage thread to avoid bad interactions with the main
thread.

The problem with the above model is that we never took into account
multi-threaded environments when threads are originated at the
application level. The EGL specification does not forbid the same
EGLDisplay to be used by different threads within the same application.
Only different contexts are required.

Therefore, given that we would use per-display event queues instead of
per-thread queues, we would make all threads share the same queues, with
higher chances of causing bad interactions between them.

This change fixes this threading issue by creating per-thread event
queues and making use of proxy wrappers.