Handle EGL_KHR_track_references
As part of our automotive efforts to support Vulkan Wayland WSI, we are
building a custom EGLStreams producer to hook into our current
libnvidia-egl-wayland.so implementation.

One of the issues encountered was that EGLDisplay would be completely
torn down when calling eglTerminate() on it without taking into account
how many entities were referencing it (e.g. when recreating the
corresponding Vulkan swapchain).

To improve this, this change adds the appropriate refcounting and
handles EGL_KHR_track_references whenever that extension is used by the
application.

Additionally, a new wlEglInitializeSurfaceExport() export has been added
to bootstrap a WlEglSurface that is not an EGLSurface producer
(EGL_KHR_stream_producer_eglsurface).

Signed-off-by: Sam Ricker <sricker@nvidia.com>