Add a capability override mechanism
libnvidia-egl-wayland.so uses one method or another to create the
underlying EGLStream for a given EGLSurface according to the set of
capabilities advertised by the server, which in turn depend on the
available driver extensions.

This change adds an environment variable to override the server
capabilities so we can force a specific stream creation method.

Signed-off-by: Ashutosh Agarwal <asagarwal@nvidia.com>