Fix several multi-threaded issues
For a given thread, multiple wayland displays can be used, concurrently
or not. Therefore, we need a list of per-display event queues for each
thread rather than a single event queue.

Additionally, there are several places where we release the global lock
while using an EGLSurface (e.g. dispatching wayland queues or in
eglSwapBuffers). Other threads might destroy the surface at that time,
so we need to re-validate it once the lock is re-acquired.