Properly handle Vulkan present modes
This is another piece in our automotive efforts to support Vulkan
Wayland WSI through libnvidia-egl-wayland.so

This change adds the required hooks and protocols for clients to be able
to setup swapchains to operate in FIFO or MAILBOX present modes.

Two new pre/post present exports have been added for proper FIFO
management.

Additionally, the protocol has been extended with a new
'wl_eglstream_controller::attach_eglstream_consumer_attribs' request so
that clients can indicate what present mode (and size of the
queue/swapchain if FIFO) they intend to use for the compositor to
properly setup its end of the underlying stream.

Signed-off-by: Sam Ricker <sricker@nvidia.com>