Use dladdr to check wl_object interface type
In a couple of places in our Wayland EGL code, we check whether
a wl_object's interface type is what we expect by performing a
simple pointer comparison with the address of the interface's symbol.

However, several libraries such as SDL2 and Qt define their own
symbols with the same name as those defined in libwayland-client.so
which can cause problems with that approach. Although these
libraries really shouldn't be doing that, we can work around the
issue by using dladdr(3) to look up the name of the symbol corresponding
to the wl_object's interface pointer and then performing a string
comparison on that.

This change implements the above functionality, and modifies the
wlEglIsWaylandWindowValid and wlEglIsWaylandDisplay functions to use it.

Signed-off-by: Erik Kurzinger <ekurzinger@nvidia.com>