Better multi-threading queue management
In a multi-threaded environment, while one thread tries to dispatch its
own queue, the destroyed queues of other thread could be accessed in
wayland API read_events if some events are pending on the destroyed
queues.

Therefore, a reference counting mechanism is added for queues. While
expecting a callback from the queue, refCount is incremented. After
receiving the callback refCount is decremented. This way we can defer
the queue destruction if it is in use. refCount is used here instead of
a busyFlag because more than one event can be expected from the same
queue.

The refCount doesn't need to be used for all the events. For example,
registry listener or display sync callbacks are received immediately
after a roundtrip. In this case we don't need to use refCount.

Only for the events similar to wl_buffer::release or wl_surface::frame
this refCount has to be incremented before assigning queue and decrement
after the callbacks are received.

A new dangling queue list is added to the display to keep the removed
queues from thread list. The dangling queue list is destroyed either
while decrementing refCount to 0 or when the display is terminated.

Signed-off-by: Maniraj D <md@nvidia.com>