Defer checking for swapinterval override until eglSwapBuffers()
Given the recent queue management refactoring in WlEglSurface, we can no
longer request swapinterval override notification at surface creation
time and process it at eglSwapBuffers() time, even though it is only
required to be done once.

This change defers the whole thing until eglSwapBuffers(), which is the
time we can successfully do it, since we require an EGLContext to be
current.

This re-introduces per-display event queues, since we no longer need
per-thread ones after surfaces are given their own, and finer-grain
locks are used instead. The result is the removal of big chunks of now
dead code.

Signed-off-by: Ashutosh Agarwal <asagarwal@nvidia.com>