egl-wayland: don't call wlEglSetError from wlEglDestroySurface
If wlEglDestroySurface is called for a surface which has already been freed (or
is otherwise invalid), it will correctly detect this and bail out immediately.
However, before doing so, it will call wlEglSetError to set an EGL_BAD_SURFACE
error.  If wlEglDestroySurface is being called from destroy_callback, though,
due to the surface's window having been destroyed, we will no longer have a
valid EGL display handle which causes the wlEglSetError call to segfault. This
can be triggered by any application that explicitly calls eglDestroySurface for
a window surface before destroying the associated window (or terminating).

To fix this, we should move the wlEglSetError call to wlEglDestroySurfaceHook
if wlEglDestroySurface returns EGL_FALSE. This ensures we will only try to set
an EGL_BAD_SURFACE error when we have a valid EGL display handle - i.e. when an
invalid surface is passed to eglDestroySurface. The only case where
wlEglDestroySurface returns false is if it can't find the given surface, so
this still aligns with the spec.

Signed-off-by: Erik Kurzinger <ekurzinger@nvidia.com>