avoid segfault in wlEglCheckInterfaceType
wlEglCheckInterfaceType calls dladdr on the given symbol address and then
passes the returned dli_sname to strcmp to check if it matches the desired
interface. However, if passed an address that doesn't correspond to any
symbol, dli_sname will be NULL, causing a segfault. Add a check to return
false in this case instead.