Add minimal wayland-drm protocol implementation
This implements just enough of the wayland-drm
protocol to report which DRM device the wayland
compositor is running on to interested clients.
It does not advertise any supported formats or
modifiers, so all attempts to create a buffer
using the DRM protocols are expected to (and do)
fail with an invalid format error. Clients are
expected to utilize some other buffer-creation
mechanism, such as the dma-buf protocol, to create
GPU-accessible buffers.

The wayland-drm.xml file defining the protocol
itself is copied from the Mesa project:

  src/egl/wayland/wayland-drm/wayland-drm.xml

At commit e74c9473592dafce4fb0005555ce9406b4200012:

  Author: Adam Jackson <ajax@redhat.com>
  Date:   Fri Aug 30 15:35:22 2019 -0400

  egl/wayland: Add FP16 format support

Which is the latest version as of this writing.