set wl_eglstream fd to -1 after closing
In wlEglCreateStreamAttribHook, if the wl_eglstream handle is a file descriptor
we close the file descriptor after creating the EGLStream.  However, we don't
set the handle to -1 after doing so. This can result in the same file
descriptor being closed a second time in destroy_wl_eglstream_resource when the
client destroys the resource.  By that point, the fd may have been recycled for
some other potentially important purpose, resulting in various kinds of errors
from corruption to crashes.

To avoid this problem, we should set wlStream->handle to -1 after closing the
file descriptor.