switch to per-EGLDisplay locking where possible
This change replaces the use of the external API lock with a per-EGLDisplay
lock in egl-wayland wherever possible, Each WlEglDisplay struct will now
have a mutexLock member, and any functions that operate solely on a single
EGLDisplay will rely on that for synchronization. It also replaces the
global wlEglSurfaceList with a per-EGLDisplay list.