dmabuf buffer release event thread
Unlike with the server-side streams surfaces,
frames are only returned to the dmabuf-based
streams after the client processes a buffer
release event corresponding to the frame's
buffer. The compositor sends buffer release
events after it has latched a new buffer attach
commit operation and is no longer using the old
buffer. However, if the client gets ahead of the
compositor by a few frames, its main thread may
be stuck in the EGL library waiting for the frame
in question to be returned. It can not process any
wayland events that would return the frame it is
waiting for in this state. Similarly, the
surface's damage thread will likely be blocked
on the stream sync object waiting for a new frame
to process, where it also can not process wayland
events. This problem is exacerbated by the current
Vulkan stream producer library, which waits for
the prior frame to be returned just after
presenting a new frame, meaning it will never get
back in to the wayland EGL client library code
between sending a new frame and waiting for the
prior frame's release event to be processed,
guaranteeing it will hit this race condition on
the second frame.

To address this, a new thread is introduced soley
to process buffer release events sent to dmabuf-
based wl_buffer objects. This thread waits for
input using poll() rather than a blocking EGL
call. Communication between the application
threads in the wayland EGL platform library and
the buffer release event processing thread goes
through a pipe() FD pair rather than shared memory
structures, meaning the poll() call can wait for
both input from the main threads and the wayland
compositor at the same time. While the limited
responsibilities of the thread eliminate the need
for most locking, each stream image (consisting of
an EGLImage/wl_buffer pair) is now mutexed as the
content must be read and updated both by the main
threads in response to EGLImage stream consumer
events, eglSwapBuffers/vkQueuePresentKHR calls,
and the surface context destructor, as well as by
the buffer release event thread when processing
buffer releases. Additionally, care must be taken
to ensure images are only added to the stream
image array once they are fully initialized
(meaning the variable tracking the number of
images in the array should only be incremented
once a fully initialized image is present in the
new slot) and the image slots should only be torn
down once the buffer release event thread has been
stopped. Note this still allows "removal" of an
EGLImage and reuse of its slot during the
lifetime of a surface as long as the removal event
locks the mutex before updating the image slot.
Only removal of the slot itself is unsafe while
the buffer release event thread is running.

The buffer release event thread must also avoid
accessing the image array (The array of pointers
to the actual images) entirely to avoid more
mutexes or a race condition between it and the
application thread(s) given the array can be
resized when images are added to the stream, which
in turn requires freeing the old array and
allocating a new one. To accomplish this, stash a
pointer to the surface in each WlEgLStreamImage
and pass the image as the user data parameter to
the buffer release callback for dmabuf-based
wl_buffers rather than passing the WlEglSurface as
the user data parameter and iterating through the
array to find the associated WlEglStreamImage in
the callback.