egl-wayland: avoid assigning internal handle to NULL
With a recent patch, in wlEglGetInternalHandleExport() API, for the
EGL_OBJECT_SURFACE_KHR type, if the internal handle is not found, it is
reset to NULL. It causes the egl external APIs not to generate proper EGL
error codes for some error cases. For an example, when an invalid
EGLSurface (some random value) is passed to eglMakeCurrent() API, the
expected error code is EGL_BAD_SURFACE, but it results in EGL_SUCCESS as
the surface pointer is being reset to NULL (EGL_NO_SURFACE) in
wlEglGetInternalHandleExport() API. This leads to failure of some of the
CTS cases.