egl-wayland: check interface name in wlEglCheckInterfaceType
Currently, wlEglCheckInterfaceType will pass the provided wl_interface
pointer to dladdr and check whether the symbol name matches the expected
interface name.  However, this assumes that the application is using the
same wayland client implementation as libnvidia-egl-wayland.so is linked
against, which may not be the case. For instance, rust has its own
implementation of the wayland client library, including all wayland
interfaces.

Instead, wlEglCheckInterfaceType should do a string comparison on the
actual name of the wl_interface.

In order to maintain the current behavior where it's safe to pass an
arbitrary pointer to this function, we must now also ensure the entire
wl_interface struct, as well as the name string, are in the process's
address space. To achieve this we replace the wlEglPointerIsDereferencable
function with a more generic wlEglMemoryIsReadable, which can check whether
an arbitrarily sized block of memory is readable. It does this by passing a
pointer to the memory to write, with the destination fd set to a dummy
pipe. As documented, write will fail with EFAULT if the provided memory is
not in the process's address space.

For simplicity's sake, all other uses of wlEglPointerIsDereferencable will
be replaced with this new function as well. This also removes the
dependency on libdl since we no longer need to call mincore, so that has
been removed from all build configuration files.