egl-wayland: don't attach NULL buffer in wlEglDestroySurface
wlEglDestroySurface will currently attach a NULL buffer to the wl_surface
backing a given WlEglSurface on destruction. This is problematic if the
WlEglSurface corresponds to a Vulkan swapchain because a new swapchain may
have already been created for that wl_surface, and hence, a new buffer
attached. Attaching a NULL buffer will trigger a xdg_wm_base configure
event for the surface, likely prompting the app to re-create the swapchain
*again*, destroy the old one, and thus get *another* configure event, ad
infinitum. In practice, this manifests as flickering as the app re-creates
its swapchain every frame.

It's not actually necessary for WlEglDestroySurface to attach a NULL buffer
to the surface. This may leave the surface contents in an undefined state,
as per the documentation for wl_surface_attach, however neither the EGL nor
Vulkan specs make any mention of what the contents of the native surface
associated with an EGLSurface or swapchain should be after destruction.