src/wayland-eglutils.c: check for EINTR and EAGAIN
In 2eb4628d64a8297f ("egl-wayland: check interface name in
wlEglCheckInterfaceType", 2021-08-30, 1.1.8), the
wlEglPointerIsDereferencable function was replaced with a new
wlEglMemoryIsReadable function.

The new wlEglMemoryIsReadable function relies on the write function
returning EFAULT when passed a buffer that is "outside your accessible
address space".

However, there's nothing to prevent that write call from being
interrupted and returning EINTR before it can complete validation
of the input arguments.

Add a loop to retry the call if it's interrupted with a signal.

Additionally, there's no guarantee that a just-created pipe will
always be able to immediately accept at least one byte without
blocking in which case an EAGAIN error will be returned.

Detect an EAGAIN error and treat that as success since the EAGAIN
error indicates that the call can succeed if retried implying the
arguments are otherwise valid.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>