src/wayland-eglutils.c: check both ends for readability
When using the write function call to check for whether or not the
buffer being written is "outside your accessible address space,"
it's important to consider any shortcuts the write function may be
taking when validating the arguments.

In particular, if the passed in length to write is zero it may not
do any validation at all.  To that end, treat a length of 0 as 1.

Furthermore, since the destination is a pipe that can only ever
accept a small number of bytes without blocking, there's no need
for the write function call to validate the entire memory range of
the buffer being passed to be written, merely the initial portion.

Work around this optimization by attempting to write a byte from
the very end of the memory range to the pipe.  Be careful to avoid
a PIPE_BUF size optimization by the write call by draining the
single byte before attempting to write the entire range.

(If a memory range of exactly PIPE_BUF bytes is passed, but there's
only room in the pipe for PIPE_BUF-1 bytes because the write of a
single byte from the end used up one byte, then the write call is
expected to return EAGAIN in which case there's really no need for
it to validate the buffer's memory address -- another optimization
shortcut to avoid.)

While this will not cause the entire memory range to be validated
in the face of write function optimizations, it will force an extra
check at the very end which provides a much more robust test than
only checking the initial portion of the memory range.

An alternative to provide even more checking would be to attempt
to write at least one byte from every PAGE_SIZE block within the
memory range being checked.  However, that would definitely incur
a performance penalty for large memory ranges and even that could
still potentially miss unwritable holes in the middle.

Adding a check of the very end of the memory range provides a speedy
compromise that will provide a more accurate answer in the face of
write function optimizations without incurring a performance penalty.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>