src/wayland-eglutils.c: treat EINVAL like EFAULT
If the write function thinks the pointer and/or length arguments
are invalid, it may return EINVAL instead of EFAULT.

In the context of wlEglMemoryIsReadable that will do just as well
as an indication that the memory is most likely unreadable.

Therefore allow an EINVAL error without causing an assertion failure.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>