wayland: Fail eglGetDisplay if wl_drm is not available
This patch does two things:
- checks if wl_drm is in use on the server, and uses it to get the name
  of the drm device the compositor is driving.
- Find an EGLDevice that matches the path returned by wl_drm.

If wl_drm and the needed extensions are not present, or if a matching
EGLDevice is not found, then we fail. Right now we only support
running on the same GPU as the compositor, so any of these being
missing means that is not the case.