wayland-egl: avoid unnecessary commit
Earlier in wlEglCreateFrameSync() API, a wl_surface_commit() was called
after a frame callback was created with wl_surface_frame(). When the frame
is actually committed with wlEglSendDamageEvent()
(either directly from SwapBuffers / from damage thread), another
wl_surface_commit() is called. So for each SwapBuffers() call, two
wl_surface_commit() are getting submitted which may result in some
performance issues as weston endup in scheduling two repaints for each
SwapBuffers() call.

To avoid this, creating frame callback only before the actual
wl_surface_commit() call (either from SwapBuffers / damage thread).

But with this update the app thread need to wait for the damage_thread to
commit the surface and generate a frame sync, introduced a conditional
variable and a mutex for this purpose.