egl-wayland: support PRIME render offload
Currently, egl-wayland only supports running EGL applications on the
same GPU as the compositor is using. With this change, if the
environment variable __NV_PRIME_RENDER_OFFLOAD is set, it will instead
always use the first NVIDIA GPU available. To ensure the compositor,
which may be running on a different, non-NVIDIA GPU, is able to import
the buffers we send it, they will be configured with a linear DRM format
modifier, which currently also implies that they will be allocated in
sysmem.

In this way, users of hybrid-graphics systems where the integrated GPU
is driving the display can still run individual applications on the
NVIDIA GPU.