vulkan-producer: proper cleanup for failure case
Earlier just the ProducerData was freed in case of any failure with
ProducerInit() call, but the related resources were not de-initialized (ex.
WlEglSurface), now make it proper by calling ProducerCleanup() in case of
any failure.

Also renamed the wlEglInitializeSurfaceExport. With recent ABI related
updates to wayland-egl, wlEglInitializeSurfaceExport() API allocates the
wlEGLSurface and initializes it with the given arguments. So the name
*Create would be more appropriate.