Defer resizes until next SwapBuffers
Applications which call wl_egl_window_resize() in the middle of a frame
from another thread will hang, as MakeCurrent() is called from the
wrong thread.

The obvious fix for this is to delay the resize until the next
eglSwapBuffers() call. An SDL-side implementation of this exists:
https://github.com/libsdl-org/SDL/pull/4821
But since some things should update immediately on resize (even if the
attached buffer doesn't), it seems better to have this in egl.

This patch fixes games such as:
- Imposter Factory:
https://store.steampowered.com/app/1182620/Impostor_Factory/
- Life is Strange:
https://store.steampowered.com/app/319630/Life_is_Strange__Episode_1/

Both would hang on startup if run against a wayland-enabled build of
SDL, but work perfectly with this patch.

Signed-off-by: David Gow <david@ingeniumdigital.com>