egl-wayland: fix eglSwapInterval with GBM compositors
In wlEglSwapIntervalHook, if the compositor does not support the
wl_eglstream_display interface we will return early without setting
surface->swapInterval. This will result in eglSwapBuffers always waiting
for the frame callback from the previous swap even if the application
has set the swap interval to 0, which is not the intended behavior.

To fix it, we should set surface->swapInterval even if wlStreamResource
is null. We should still only set surface->pendingSwapIntervalUpdate if
wlStreamResource is not null, though.