Resize immediately if surface is current.
Commit ddaa272 ("Defer resizes until next SwapBuffers") was perhaps
overzealous, as most applications render and resize on the same thread.
Given the resize operation destroys the current surface contents,
deferring it to just before swapbuffers broke rendering for the first
frame after a resize (see issue #57).

Instead, detect if the surface being resized is bound to the current
thread, and only defer the resize if it is not. This is still not
technically correct (ideally we'd defer resizing the buffer until the
next "safe point", when its contents are already being discarded, e.g.
after a glClear()), but it preserves the old behaviour in the cases
where it worked, but still fixes games which render from a different
thread to the one which handles resizes. (I've tested this with Impostor
Factory and Life is Strange, and both work, albeit still losing frame
contents during resizes.)

This was suggested by @erik-kz in #57.