EGL/Wayland: Fix matching EGL_DEVICE_EXT attributes in existing displays
In the EGL Wayland library, record the EGL_DEVICE_EXT attribute in the
WlEglDisplayRec struct.

In wlEglGetPlatformDisplayExport, record the EGL_DEVICE_EXT attribute in
a new WlEglDisplayRec::requestedDevice field, and check that the device
matches when searching for an existing EGLDisplay.

Previously, two calls to eglGetPlatformDisplay with the same wl_display
pointer but different EGL_DEVICE_EXT attributes would use return the
same EGLDisplay.