EGL/Wayland: Disable NV -> NV offloading with EGL_EXT_explicit_device
Disallow offloading between NVIDIA devices for Wayland, because it's
currently broken in both Weston and Mutter.

With Weston, it doesn't work because Weston will try to use the image
for overlay or scanout, which doesn't work with a pitch linear sysmem
image due to bug 3770050.

Mutter doesn't correctly handle external-only images, so it'll try to
import the EGLImage into a GL_TEXTURE_2D texture, which also fails.