egl-wayland: Expose the functionality of the presentation-time protocol
The following functions are added and exposed to the users of this
library:

wlEglPostPresentExport2: An updated version of wlEglPostPresentExport
that expects a pointer to a __GLNVpresentEventInfoRec structure. If the
pointer is not NULL and the compositor supports the presentation-time
extension, a presentation-time feedback request is sent along with the
presentation request and the pointer is passed as a parameter to the
callback functions that process the feedback events. In the callback
functions, the status field of these structures are updated to indicate
that the requested presentation has been performed.

wlEglProcessPresentationFeedbacksExport: A non-blocking function to
dispatch all received presentation feedback events on demand. Returns
the number of updated presentation event information structures.

wlEglWaitAllPresentationFeedbacksExport: A blocking function that waits
for all events corresponding to the sent feedback requests to be
received. Designed to be explicitly called before destroying
WlEglSurface objects to prevent events to be placed to a destroyed event
queue. There are several reasons for this operation to be implemented as
a separate function that is exposed to the users:
  * It is possible for the clients of this library to destroy the
    resources passed to wlEglPostPresentExport2 function before they
    destroy the corresponding WlEglSurface.
  * Clients of this library could use a counter that counts the number
    of presentation requests sent and the return value of the
    wlEglProcessPresentationFeedbacksExport function to wait for all in
    flight presentation feedback requests. However, they are allowed to
    change the number of presentation requests they expect by choosing to
    ignore some results, creating a disparity between the number of in
    flight requests and expected number of updates.