wayland: Fix dmabuf protocol registration
Currently we are registering our listener for the dmabuf protocol
in dmabuf_set_interface. This is called when we make a new registry
on a wl_proxy wrapper for our wl_display in wlEglInitializeHook.

The issue is that dmabuf_handle_modifier is never called. The
compositor will send the event, but the client library will never
deliver it to us even though we have bound the protocol and registered
a listener. I'm not sure exactly why this is, but it has something
to do with the proxy wrapper we are using. Other handlers, such as
for eglstreams, bind the registry but defer adding a listener until
after we have finished dealing with the wrapper.

This change makes linux_dmabuf add a listener after the registry
globals have been enumerated. This makes dmabuf_handle_modifier
finally start getting called and populating our drm format list.

It also actually passes the modifiers we received from the compositor
to streamImageConsumerConnect. This is needed for future work on
dynamic modifier renegotiation.