wayland: Add dmabuf format set
This adds helper functions for handling sets of dmabuf formats and
their associated modifiers. This will be needed for later dmabuf
v4 feedback structs, where we have multiple places to handle lists
of formats and want to be able to share code.

This adds the WlEglDmaBufFormatSet struct, and a couple functions
for easily interacting with it. This is mostly all existing
functionality from our current dmabuf protocol implementation, just
wrapped up in a reusable struct.