wayland: Realloc surface if the compositor sends updated feedback
In the case of on-the-fly negotiation with a compositor that supports
dmabuf v4, we may get an updated set of modifiers that we need to
react to. This turns wlEglResizeSurfaceIfRequired into
wlEglReallocSurfaceIfRequired, which checks if we need to reallocate
the surface and does so.

Note that this functionality is really for per-surface feedback objects,
where the compositor can resend scanout tranches if the surface is
fullscreen. This CL adds the ability to realloc before we enable
per-surface feedback, since per-surface feedback tells the compositor
we support dynamically reallocating the surface to follow what it
sends.