wayland: Register per-surface feedback
This adds the final piece of dynamic modifier renegotiation on wayland.
This change gets a feedback object for the surface that the compositor
will use to send us updates on supported formats/modifiers.

This is important for enabling features such as direct dGPU scanout
on wayland. Currently we use a linear modifier when running in PRIME, but
with this change we can also react to the compositor sending us dGPU
scanout modifiers if we are fullscreened.