wayland: Only search dmabuf format sets if surface has an eglConfig
This fixes a vulkan wsi regression on wayland introduced by CL 32234950.
Vulkan surfaces will not have a defined eglConfig, and therefore we
cannot check what format they would like to use.

Eventually we need to fix this and plumb modifier selection through on
vulkan, however given the way we currently sit vulkan on top of EGL this
will not be a simple fix. This change is intended as a short-term
bandaid to fix the regression while allowing GL/EGL to continue using
dmabuf v4.