wayland: Use linear modifier for prime when appropriate
Originally we removed forcing the linear modifier when running prime
since we could count on finding the linear modifier somewhere in the
list of intel/AMD modifiers. However with dmabuf v4 we will look for
a tranche of modifiers for a specific device (the NV one) and use the
modifiers for that. A non-fullscreen app in a wayland prime setup may
only get one tranche for the main device, meaning it finds no modifiers
to initialize the consumer with.

This change goes back to explicitly specifying the linear modifier for
prime. This happens with prime enabled and when:
1. We are on an older protocol version (< 4). This is done just to be
   explicit.
2. We could not find a format set for this device.