egl-wayland: set unprocessedFeedback flag
A prior commit added logic to reallocate window surfaces after receiving
a new dma-buf feedback from the compositor. This keys off of the
unprocessedFeedback flag, however currently that flag never actually
gets set. This change sets the flag after receiving a new feedback, so
that we will reallocate the surface on the next eglSwapBuffers call.