egl-wayland: handle multiple tranches for the same device
To determine the set of format modifiers to use when creating a surface
we will find the first tranche for our DRM device, or else the
compositor's main device, and then look for our DRM format in that
tranche's list of formats to find the modifier list. If our format is
not in the list, then we won't use explicit modifiers when creating the
surface.

However, some compositors will advertise multiple tranches for a given
DRM device with a different set of formats in each one. Kwin is one
example. In that case, the tranche we select above might not contain our
format, but it could still be in one of the other tranches for the same
device.

Instead of simply picking the first tranche for the desired device, with
this change we will loop over the format list of each of that device's
tranches until we find a tranche that contains the format we want.