egl-wayland: destroy all proxies before destroying their queues
The latest version of libwayland-client.so will print a warning if a
queue is destroyed while a proxy is still assigned to it. This warning
is triggered during EGL termination by our linux_dmabuf object and the
default dmabuf feedback object for the display. It is also triggered
during EGLSurface destruction by the surface's dmabuf feedback object.
That last object never gets destroyed at all, in fact.

This change ensures that the above objects are all destroyed *before*
their queues.